[proof] Make LeafCount type alias
Add a type alias for LeafCount as u64 so that 32-bit builds will
continue to work.  This is required for external HSM code that
needs to include types crate.  We lose some type checking but
gain readability and consistency on what is a Version versus
what is a LeafCount.

+++ storage/libradb/src/event_store/mod.rs
@@ -75,7 +75,7 @@ impl EventStore {
         let mut iter = self.db.iter::<EventSchema>(ReadOptions::default())?;
         iter.seek_for_prev(&(version + 1))?;
         let num_events = match iter.next().transpose()? {
-            Some(((ver, index), _)) if ver == version => (index + 1) as usize,
+            Some(((ver, index), _)) if ver == version => (index + 1),
             _ => unreachable!(), // since we've already got at least one event above
         };
