+++ storage/libradb/src/ledger_store/mod.rs
@@ -128,7 +128,7 @@ impl LedgerStore {
     ) -> Result<AccumulatorProof> {
         Accumulator::get_proof(
             self,
-            (ledger_version + 1) as usize, /* num_leaves */
+            ledger_version + 1, /* num_leaves */
             version,
         )
     }
@@ -151,7 +151,7 @@ impl LedgerStore {
         let txn_hashes: Vec<HashValue> = txn_infos.iter().map(TransactionInfo::hash).collect();
         let (root_hash, writes) = Accumulator::append(
             self,
-            first_version as usize, /* num_existing_leaves */
+            first_version, /* num_existing_leaves */
             &txn_hashes,
         )?;
         writes
@@ -175,7 +175,7 @@ impl LedgerStore {
 
     /// From left to right, get frozen subtree root hashes of the transaction accumulator.
     pub fn get_ledger_frozen_subtree_hashes(&self, version: Version) -> Result<Vec<HashValue>> {
-        Accumulator::get_frozen_subtree_hashes(self, version as usize + 1)
+        Accumulator::get_frozen_subtree_hashes(self, version + 1)
     }
 }
